Restore error_values as field
PeakAnalyst and ResultsContainer had both had the field error_values. We'd taken
it out in order to use an external function to parse them from the GRASS parser,
which reduced redundancy. However, both objects kept on looking for
self.error_values, so I gave in and gave it back to them. However, the
error_values are still set using the function parse_error_values().